+++ behave/parser.py
@@ -1,8 +1,8 @@
 # -*- coding: UTF-8 -*-
 
 from __future__ import absolute_import, with_statement
-import six
 import re
+import six
 from behave import model, i18n
 from behave.textutil import text as _text
 
@@ -195,11 +195,14 @@ class Parser(object):
             raise ParserError(message, self.line, self.filename, line)
         name = line[len(keyword) + 1:].strip()
         self.examples = model.Examples(self.filename, self.line,
-                                       keyword, name)
+                                       keyword, name, tags=self.tags)
         # pylint: disable=E1103
         #   E1103   Instance of "Background" has no "examples" member
         #           (but some types could not be inferred).
         self.statement.examples.append(self.examples)
+
+        # -- RESET STATE:
+        self.tags = []
 
 
     def diagnose_feature_usage_error(self):
@@ -297,10 +300,39 @@ class Parser(object):
             return True
         return False
 
-    def subaction_detect_next_scenario(self, line):
+    # def subaction_detect_next_scenario(self, line):
+    #     if line.startswith("@"):
+    #         self.tags.extend(self.parse_tags(line))
+    #         self.state = "next_scenario"
+    #         return True
+    #
+    #     scenario_kwd = self.match_keyword("scenario", line)
+    #     if scenario_kwd:
+    #         self._build_scenario_statement(scenario_kwd, line)
+    #         self.state = "scenario"
+    #         return True
+    #
+    #     scenario_outline_kwd = self.match_keyword("scenario_outline", line)
+    #     if scenario_outline_kwd:
+    #         self._build_scenario_outline_statement(scenario_outline_kwd, line)
+    #         self.state = "scenario"
+    #         return True
+    #
+    #     # -- OTHERWISE:
+    #     return False
+
+    def subaction_detect_taggable_statement(self, line):
+        """Subaction is used after first tag line is detected.
+        Additional lines with tags or taggable_statement follow.
+
+        Taggable statements (excluding Feature) are:
+        * Scenario
+        * ScenarioOutline
+        * Examples (within ScenarioOutline)
+        """
         if line.startswith("@"):
             self.tags.extend(self.parse_tags(line))
-            self.state = "next_scenario"
+            self.state = "taggable_statement"
             return True
 
         scenario_kwd = self.match_keyword("scenario", line)
@@ -315,12 +347,20 @@ class Parser(object):
             self.state = "scenario"
             return True
 
+        examples_kwd = self.match_keyword("examples", line)
+        if examples_kwd:
+            self._build_examples(examples_kwd, line)
+            self.state = "table"
+            return True
+
         # -- OTHERWISE:
         return False
 
     def action_feature(self, line):
         line = line.strip()
-        if self.subaction_detect_next_scenario(line):
+        # OLD: if self.subaction_detect_next_scenario(line):
+        if self.subaction_detect_taggable_statement(line):
+            # -- DETECTED: Next Scenario, ScenarioOutline (or tags)
             return True
 
         background_kwd = self.match_keyword("background", line)
@@ -332,12 +372,28 @@ class Parser(object):
         self.feature.description.append(line)
         return True
 
-    def action_next_scenario(self, line):
-        """
-        Entered after first tag for Scenario/ScenarioOutline is detected.
+    # def action_next_scenario(self, line):
+    #     """
+    #     Entered after first tag for Scenario/ScenarioOutline is detected.
+    #     """
+    #     line = line.strip()
+    #     if self.subaction_detect_next_scenario(line):
+    #         return True
+    #
+    #     return False
+
+    def action_taggable_statement(self, line):
+        """Entered after first tag for Scenario/ScenarioOutline or
+        Examples is detected (= taggable_statement except Feature).
+
+        Taggable statements (excluding Feature) are:
+          * Scenario
+          * ScenarioOutline
+          * Examples (within ScenarioOutline)
         """
         line = line.strip()
-        if self.subaction_detect_next_scenario(line):
+        if self.subaction_detect_taggable_statement(line):
+            # -- DETECTED: Next Scenario, ScenarioOutline or Examples (or tags)
             return True
 
         return False
@@ -362,7 +418,9 @@ class Parser(object):
         # -- CASE: Detect next Scenario/ScenarioOutline
         #   * Scenario with scenario description, but without steps.
         #   * Title-only scenario without scenario description and steps.
-        if self.subaction_detect_next_scenario(line):
+        # OLD: if self.subaction_detect_next_scenario(line):
+        if self.subaction_detect_taggable_statement(line):
+            # -- DETECTED: Next Scenario, ScenarioOutline (or tags)
             return True
 
         # -- OTHERWISE: Add scenario description line.
@@ -382,7 +440,7 @@ class Parser(object):
           * examples for a ScenarioOutline, after ScenarioOutline steps
 
         DETECT:
-          * next Scenario/ScenarioOutline
+          * next Scenario/ScenarioOutline or Examples (in a ScenarioOutline)
         """
         # pylint: disable=R0911
         #   R0911   Too many return statements (8/6)
@@ -400,13 +458,8 @@ class Parser(object):
             self.statement.steps.append(step)
             return True
 
-        if self.subaction_detect_next_scenario(line):
-            return True
-
-        examples_kwd = self.match_keyword("examples", line)
-        if examples_kwd:
-            self._build_examples(examples_kwd, line)
-            self.state = "table"
+        if self.subaction_detect_taggable_statement(line):
+            # -- DETECTED: Next Scenario, ScenarioOutline or Examples (or tags)
             return True
 
         if line.startswith("|"):